backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Retry logic
tenacity>=8.2.0

# Gemini AI (Phase 2)
google-generativeai>=0.3.0

# Future dependencies (Phase 3+)
# redis>=5.0.0  # Caching
# prometheus-client>=0.19.0  # Metrics
# slowapi>=0.1.9  # Rate limiting
